Item update process refactored
The item update process has been refactored into a series of
implimentations of an abstract EndOfDayItemUpdater class. These
instances separate the SellIn and the Quality value updates of the Item.
Each instance is a realization of each item category that determins how
an Item's Quality will be updated. This will allow for easier
modification of already defined categories along with easy addition of
new Item categories.